requirements.txt
async-timeout==5.0.1
joblib==1.5.3
numpy==2.2.6
redis==7.4.0
scikit-learn==1.7.2
scipy==1.15.3
threadpoolctl==3.6.0
python-dotenv==1.2.2